backend/requirements.txt
# PDF Annotation Platform - Phase 1 Requirements
# Core dependencies for immediate development

# ========================================
# CORE BACKEND FRAMEWORK
# ========================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# ========================================
# DATABASE & ORM
# ========================================
# PostgreSQL
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# SQLAlchemy 2.0 with async support
sqlalchemy[asyncio]>=2.0.20
alembic>=1.13.0

# ========================================
# REDIS & CACHING
# ========================================
redis>=5.0.0

# ========================================
# AUTHENTICATION & SECURITY
# ========================================
# OAuth2 & JWT - compatible versions
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# Environment management
python-dotenv>=1.0.0

# ========================================
# DATA VALIDATION & SERIALIZATION
# ========================================
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# ========================================
# HTTP CLIENT & ASYNC
# ========================================
httpx>=0.25.0
aiofiles>=23.2.0

# ========================================
# TESTING FRAMEWORK
# ========================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# ========================================
# DEVELOPMENT TOOLS
# ========================================
# Code formatting & linting
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0

# ========================================
# UTILITIES
# ========================================
# JSON handling
orjson>=3.9.0

# Rich output
rich>=13.7.0

# Date utilities
python-dateutil>=2.8.2

# ========================================
# BASIC MONITORING
# ========================================
# Structured logging
structlog>=23.2.0
# System metrics
psutil

